Fixes elasticsearch wiring when using AWS and ES_HOSTS
Previously, when `ES_HOSTS` was used for an Amazon Elasticsearch domain,
the following error would result:

```
There is a circular dependency between 4 beans in the application context:
    - zipkinHttpCollector defined in file [/zipkin/BOOT-INF/classes/zipkin/server/ZipkinHttpCollector.class]
    - zipkin.autoconfigure.storage.elasticsearch.ZipkinElasticsearchStorageAutoConfiguration (field zipkin.storage.elasticsearch.InternalElasticsearchClient$Builder zipkin.autoconfigure.storage.elasticsearch.ZipkinElasticsearchStorageAutoConfiguration.clientBuilder)
    - zipkin.autoconfigure.storage.elasticsearch.http.ZipkinElasticsearchHttpStorageAutoConfiguration (field java.util.List zipkin.autoconfigure.storage.elasticsearch.http.ZipkinElasticsearchHttpStorageAutoConfiguration.networkInterceptors)
    - zipkin.autoconfigure.storage.elasticsearch.aws.ZipkinElasticsearchAwsStorageAutoConfiguration (field java.util.List zipkin.autoconfigure.storage.elasticsearch.http.ZipkinElasticsearchHttpStorageAutoConfiguration.networkInterceptors)
    - awsSignatureVersion4
```

This fixes that by unwinding an initialization concern around the OkHttp
client used.
