+++ zipkin-autoconfigure/storage-elasticsearch-http/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,4 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+zipkin.autoconfigure.storage.elasticsearch.http.ZipkinElasticsearchOkHttpAutoConfiguration,\
 zipkin.autoconfigure.storage.elasticsearch.http.ZipkinElasticsearchHttpStorageAutoConfiguration,\
 zipkin.autoconfigure.storage.elasticsearch.http.brave.TraceZipkinElasticsearchHttpStorageAutoConfiguration
